Rough team head to head, team images, lots of fixes
The browser has been improved and new endpoints added to get team head
to heads as well as images for the squad members on a team.

This is built against the SNAPSHOT release of the pa-football-client
package I have locally. After that is released this will be properly
built against a stable pa client jar.